requirements.txt
absl-py==1.0.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
dill==0.3.5.1
Django==4.0.4
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.3
h5py==3.7.0
idna==3.3
image==1.5.33
importlib-metadata==4.11.4
importlib-resources==5.7.1
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-utils==1.0.13
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
promise==2.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.15.0
sqlparse==0.4.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metadata==1.8.0
tensorflow-metal==0.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
